Secured some functions with token id, minor ui changes

+++ ui/src/app/components/timesheet-details/timesheet-details.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { TimesheetsService } from 'src/app/services/timesheets.service';
 import { ActivatedRoute, Router } from '@angular/router';
+import {IdentifierService} from "src/app/services/identifier.service"
 
 @Component({
   selector: 'app-timesheet-details',
@@ -12,7 +13,7 @@ import { ActivatedRoute, Router } from '@angular/router';
     <h1>Timesheet</h1>
     <form>
       <div class="form-group">
-        <label for="Month">Date</label>
+        <label for="Month">Date:</label>
         <input
           type="text"
           class="form-control"
@@ -22,7 +23,7 @@ import { ActivatedRoute, Router } from '@angular/router';
         />
       </div>
       <div class="form-group">
-        <label for="TimeIn">Time In</label>
+        <label for="TimeIn">Time In:</label>
         <input
           type="text"
           class="form-control"
@@ -32,7 +33,7 @@ import { ActivatedRoute, Router } from '@angular/router';
         />
       </div>
       <div class="form-group">
-        <label for="TimeOut">Time Out</label>
+        <label for="TimeOut">Time Out:</label>
         <input
           type="text"
           class="form-control"
@@ -42,7 +43,7 @@ import { ActivatedRoute, Router } from '@angular/router';
         />
       </div>
       <div class="form-group">
-      <label for="NumHours">Hours</label>
+      <label for="NumHours">Hours:</label>
       <input
         type="text"
         class="form-control"
@@ -74,21 +75,24 @@ import { ActivatedRoute, Router } from '@angular/router';
 </div>  
 `,
   styles: [
-    'input {color: rgb(6, 134, 128);}'
+    'input {color: rgb(6, 134, 128);}',
+    'label {color: rgb(6, 134, 128); font-weight: bold; margin-right: 5px;}'
   ]
 })
 export class TimesheetDetailsComponent implements OnInit {
   currentTimesheet:any = null;
   message = '';
+  userid = this.identifier.getIdentifier();
 
   constructor(
-    private timesheetService: TimesheetsService,
+    private timesheetService: TimesheetsService, public identifier: IdentifierService,
     private route: ActivatedRoute,
     private router: Router) { }
 
   ngOnInit(): void {
     this.message = '';
     this.getTimesheet(this.route.snapshot.paramMap.get('id'));
+    this.userid = this.identifier.getIdentifier()
   }
 
   getTimesheet(id:any): void {
@@ -96,7 +100,6 @@ export class TimesheetDetailsComponent implements OnInit {
       .subscribe(
         data => {
           this.currentTimesheet = data;
-          //console.log(data);
         },
         error => {
           console.log(error);
@@ -114,7 +117,6 @@ export class TimesheetDetailsComponent implements OnInit {
       .subscribe(
         response => {
           this.currentTimesheet.current = status;
-          //console.log(response);
         },
         error => {
           console.log(error);
@@ -122,26 +124,32 @@ export class TimesheetDetailsComponent implements OnInit {
   }
 
   updateTimesheet(): void {
-    this.timesheetService.update(this.currentTimesheet._id, this.currentTimesheet)
-      .subscribe(
-        response => {
-          //console.log(response);
-          this.message = 'The timesheet was updated successfully!';
-        },
-        error => {
-          console.log(error);
-        });
+    if (this.currentTimesheet.EmpName == this.userid) { //Check token id - don't let request through if it's incorrect
+      this.timesheetService.update(this.currentTimesheet._id, this.currentTimesheet)
+        .subscribe(
+          response => {
+            this.message = 'The timesheet was updated successfully!';
+          },
+          error => {
+            console.log(error);
+          });
+    } else {
+      this.message = "Error: Not Authorized"
+    }
   }
 
   deleteTimesheet(): void {
-    this.timesheetService.delete(this.currentTimesheet._id)
-      .subscribe(
-        response => {
-          //console.log(response);
-          this.router.navigate(['/timesheets']);
-        },
-        error => {
-          console.log(error);
-        });
+    if (this.currentTimesheet.EmpName == this.userid) { //Check token id - don't let request through if it's incorrect
+      this.timesheetService.delete(this.currentTimesheet._id)
+        .subscribe(
+          response => {
+            this.router.navigate(['/timesheets']);
+          },
+          error => {
+            console.log(error);
+          });
+      } else {
+        this.message = "Error: Not Authorized"
+      }
   }
 }
